requirements.txt
alembic==1.17.1
blinker==1.9.0
click==8.3.0
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.2
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
gunicorn==23.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
loguru==0.7.3
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
python-dotenv==1.1.1
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
WTForms==3.2.1
